# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion de projet",
        "version": "1.0.0"
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connecte l'utilisateur au système",
                "description": "Connexion de l'utilisateur au système",
                "operationId": "loginUser",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "L'email de l'utilisateur pour la connexion",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "john@gmail.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Le mot de passe de connexion",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "123456"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "description": "Un jeton",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Le type de token",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "La durée d'expiration du token",
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6MjAyM1wvYXBpXC9sb2dpbiIsImlhdCI6MTYxOTExNTU1OCwiZXhwIjoxNjE5MjAxOTU4LCJuYmYiOjE2MTkxMTU1NTgsImp0aSI6Ilp4N0VTYjNwZGVOQW9oUFgiLCJzdWIiOjEsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.X7wwPIsEzEVFkB1R69y5-harbqjqki5Yon6DyKZCOOc",
                                        "token_type": "bearer",
                                        "expires_in": 86400
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Informations d'identification invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message d'échec d'authentification",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Ces identifiants ne correspondent pas à nos enregistrements."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Adresse de courriel non vérifiée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message d'échec d'authentification",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "Vous devez d'abord vérifier votre adresse e-mail."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/register": {
            "post": {
                "tags": [
                    "Inscription"
                ],
                "summary": "Enregistrer un nouvel utilisateur",
                "description": "Enregistrer un nouvel utilisateur",
                "operationId": "registerUser",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Le nom de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "John Doe"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "L'email de l'utilisateur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "john@gmail.com"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Mot de passe",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password",
                            "example": "Mija6!"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Confirmation du mot de passe",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "password",
                            "example": "Mija6!"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Utilisateur enregistré avec succes"
                    },
                    "422": {
                        "description": "Validation errors"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}